Update on pull request Yanina
Hi Yanina! I've checked your updates and I love the idea and setup in general! Now you have specific Flashtalking related code and also files. I would like to keep it more generic so that we can also easily expand on other platforms when needed without it becoming very specific code for each platform.   So hereby I share with you my suggestion. Let me know if you like it.
in index.js I've removed the subtype question and put it in the general type, so that we don't have too much subquestions and to keep only one platformChoices file.
In generic.js I have then removed subtype as it is already in the platformChoice and also the platformWithSubtypeInputPath as that is then also not needed anymore.
instead of FTSubtypeChoices.js I've extended PlatformChoices to keep it more in one place. If we are going to create subtypes for each platform it can become a bit hard to see what belongs to what I believe.